requirements.txt
dotenv==0.9.9
faiss-cpu==1.12.0
fastapi==0.121.1
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
openai==2.7.2
orjson==3.11.4
prompt_toolkit==3.0.52
python-dotenv==1.2.1
regex==2025.11.3
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tornado==6.5.2
typing_extensions==4.15.0
uvicorn==0.38.0